model-containers/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
google-generativeai==0.3.1
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.3.0
requests==2.31.0
